Moved banner out of the folder.
